requirements.txt
streamlit
tensorflow
numpy
pillow
scipy
scikit-learn
matplotlib
tf-keras